refactor bulk of btc/doge into base adapter
