+++ packages/chain-adapters/src/utxo/UTXOBaseAdapter.ts
@@ -1,5 +1,14 @@
 import { AssetId, ChainId } from '@shapeshiftoss/caip'
-import { bip32ToAddressNList, HDWallet, PublicKey, supportsBTC } from '@shapeshiftoss/hdwallet-core'
+import {
+  bip32ToAddressNList,
+  BTCOutputAddressType,
+  BTCOutputScriptType,
+  BTCSignTxInput,
+  BTCSignTxOutput,
+  HDWallet,
+  PublicKey,
+  supportsBTC
+} from '@shapeshiftoss/hdwallet-core'
 import { BIP44Params, KnownChainIds, UtxoAccountType } from '@shapeshiftoss/types'
 import * as unchained from '@shapeshiftoss/unchained-client'
 import WAValidator from 'multicoin-address-validator'
@@ -11,6 +20,7 @@ import {
   BuildSendTxInput,
   ChainTxType,
   FeeDataEstimate,
+  FeeDataKey,
   GetFeeDataInput,
   SignTxInput,
   SubscribeError,
@@ -32,8 +42,9 @@ import {
 } from '../utils'
 import { bnOrZero } from '../utils/bignumber'
 import { GetAddressInput } from './types'
+import { utxoSelect } from './utxoSelect'
 
-export type UTXOChainIds = KnownChainIds.BitcoinMainnet | KnownChainIds.DogecoinMainnet // to be extended in the future to include other UTXOs
+export type UtxoChainId = KnownChainIds.BitcoinMainnet | KnownChainIds.DogecoinMainnet
 
 /**
  * Currently, we don't have a generic interact for UTXO providers, but will in the future.
@@ -46,7 +57,7 @@ export interface ChainAdapterArgs {
     ws: unchained.ws.Client<unchained.bitcoin.BitcoinTx>
   }
   coinName: string
-  chainId?: ChainId
+  chainId?: UtxoChainId
 }
 
 /**
@@ -55,8 +66,8 @@ export interface ChainAdapterArgs {
  *
  * `export type UTXOChainIds = KnownChainIds.BitcoinMainnet | KnownChainIds.Litecoin`
  */
-export abstract class UTXOBaseAdapter<T extends UTXOChainIds> implements IChainAdapter<T> {
-  protected chainId: ChainId
+export abstract class UTXOBaseAdapter<T extends UtxoChainId> implements IChainAdapter<T> {
+  protected chainId: UtxoChainId
   protected assetId: AssetId
   protected coinName: string
   protected accountAddresses: Record<string, Array<string>> = {}
@@ -71,13 +82,8 @@ export abstract class UTXOBaseAdapter<T extends UTXOChainIds> implements IChainA
   }
 
   /* Abstract Methods */
-
-  abstract subscribeTxs(
-    input: SubscribeTxsInput,
-    onMessage: (msg: Transaction<T>) => void,
-    onError?: (err: SubscribeError) => void
-  ): Promise<void>
-  abstract unsubscribeTxs(input?: SubscribeTxsInput): void
+  abstract getAssetId(): AssetId
+  abstract getChainId(): ChainId
   abstract closeTxs(): void
   abstract getType(): T
   abstract getSupportedAccountTypes(): UtxoAccountType[]
@@ -85,27 +91,11 @@ export abstract class UTXOBaseAdapter<T extends UTXOChainIds> implements IChainA
   abstract getDefaultAccountType(): UtxoAccountType
   abstract getTransactionParser(): unchained.bitcoin.TransactionParser
   abstract getFeeAssetId(): AssetId
-
+  abstract accountTypeToOutputScriptType(accountType: UtxoAccountType): BTCOutputScriptType
   abstract buildBIP44Params(params: Partial<BIP44Params>): BIP44Params
-
-  abstract buildSendTransaction(tx: BuildSendTxInput<T>): Promise<{ txToSign: ChainTxType<T> }>
-
-  abstract getFeeData(input: Partial<GetFeeDataInput<T>>): Promise<FeeDataEstimate<T>>
-
-  abstract signTransaction(signTxInput: SignTxInput<ChainTxType<T>>): Promise<string>
-
   abstract getDisplayName(): string
 
   /* public methods */
-
-  getChainId(): ChainId {
-    return this.chainId
-  }
-
-  getAssetId(): AssetId {
-    return this.assetId
-  }
-
   async getAccount(pubkey: string): Promise<Account<T>> {
     if (!pubkey) {
       return ErrorHandler('UTXOBaseAdapter: pubkey parameter is not defined')
@@ -177,6 +167,190 @@ export abstract class UTXOBaseAdapter<T extends UTXOChainIds> implements IChainA
     })
     if (!btcAddress) throw new Error('UTXOBaseAdapter: no address available from wallet')
     return btcAddress
+  }
+
+  async buildSendTransaction(tx: BuildSendTxInput<T>): Promise<{
+    txToSign: ChainTxType<T>
+  }> {
+    try {
+      const {
+        value,
+        to,
+        wallet,
+        chainSpecific: { satoshiPerByte, accountType, opReturnData },
+        sendMax = false
+      } = tx
+
+      if (!value || !to) {
+        throw new Error('DogecoinChainAdapter: (to and value) are required')
+      }
+
+      let { bip44Params } = tx
+      if (!bip44Params) {
+        bip44Params = this.getDefaultBip44Params()
+      }
+
+      const path = toRootDerivationPath(bip44Params)
+      const pubkey = await this.getPublicKey(wallet, bip44Params, accountType)
+      const { data: utxos } = await this.providers.http.getUtxos({
+        pubkey: pubkey.xpub
+      })
+
+      if (!supportsBTC(wallet))
+        throw new Error(
+          'DogecoinChainAdapter: signTransaction wallet does not support signing btc txs'
+        )
+
+      const account = await this.getAccount(pubkey.xpub)
+
+      const coinSelectResult = utxoSelect({
+        utxos,
+        to,
+        satoshiPerByte,
+        sendMax,
+        value,
+        opReturnData
+      })
+
+      if (!coinSelectResult || !coinSelectResult.inputs || !coinSelectResult.outputs) {
+        throw new Error(`DogecoinChainAdapter: coinSelect didn't select coins`)
+      }
+
+      const { inputs, outputs } = coinSelectResult
+
+      const signTxInputs: BTCSignTxInput[] = []
+      for (const input of inputs) {
+        if (!input.path) continue
+        const getTransactionResponse = await this.providers.http.getTransaction({
+          txid: input.txid
+        })
+        const inputTx = getTransactionResponse.data
+
+        signTxInputs.push({
+          addressNList: bip32ToAddressNList(input.path),
+          // https://github.com/shapeshift/hdwallet/issues/362
+          // eslint-disable-next-line @typescript-eslint/no-explicit-any
+          scriptType: accountTypeToScriptType[accountType] as any,
+          amount: String(input.value),
+          vout: input.vout,
+          txid: input.txid,
+          hex: inputTx.hex
+        })
+      }
+
+      const signTxOutputs: BTCSignTxOutput[] = outputs.map((out) => {
+        const amount = String(out.value)
+        if (!out.address) {
+          return {
+            addressType: BTCOutputAddressType.Change,
+            amount,
+            addressNList: bip32ToAddressNList(
+              `${path}/1/${String(account.chainSpecific.nextChangeAddressIndex)}`
+            ),
+            scriptType: this.accountTypeToOutputScriptType(accountType),
+            isChange: true
+          }
+        } else {
+          return {
+            addressType: BTCOutputAddressType.Spend,
+            amount,
+            address: out.address
+          }
+        }
+      })
+
+      const txToSign = {
+        coin: this.coinName,
+        inputs: signTxInputs,
+        outputs: signTxOutputs,
+        opReturnData
+      } as ChainTxType<T>
+      return { txToSign }
+    } catch (err) {
+      return ErrorHandler(err)
+    }
+  }
+
+  async getFeeData({
+    to,
+    value,
+    chainSpecific: { pubkey, opReturnData },
+    sendMax = false
+  }: GetFeeDataInput<T>): Promise<FeeDataEstimate<T>> {
+    const feeData = await this.providers.http.getNetworkFees()
+
+    if (!to || !value || !pubkey) throw new Error('to, from, value and pubkey are required')
+    if (!feeData.data.average?.satsPerKiloByte || !feeData.data.slow?.satsPerKiloByte) {
+      throw new Error('undefined fee')
+    }
+
+    if (!feeData.data.fast?.satsPerKiloByte || feeData.data.fast?.satsPerKiloByte < 0) {
+      feeData.data.fast = feeData.data.average
+    }
+    // We have to round because coinselect library uses sats per byte which cant be decimals
+    const fastPerByte = String(Math.round(feeData.data.fast.satsPerKiloByte / 1024))
+    const averagePerByte = String(Math.round(feeData.data.average.satsPerKiloByte / 1024))
+    const slowPerByte = String(Math.round(feeData.data.slow.satsPerKiloByte / 1024))
+
+    const { data: utxos } = await this.providers.http.getUtxos({
+      pubkey
+    })
+
+    const utxoSelectInput = {
+      to,
+      value,
+      opReturnData,
+      utxos,
+      sendMax
+    }
+
+    const { fee: fastFee } = utxoSelect({
+      ...utxoSelectInput,
+      satoshiPerByte: fastPerByte
+    })
+    const { fee: averageFee } = utxoSelect({
+      ...utxoSelectInput,
+      satoshiPerByte: averagePerByte
+    })
+    const { fee: slowFee } = utxoSelect({
+      ...utxoSelectInput,
+      satoshiPerByte: slowPerByte
+    })
+
+    return {
+      [FeeDataKey.Fast]: {
+        txFee: String(fastFee),
+        chainSpecific: {
+          satoshiPerByte: fastPerByte
+        }
+      },
+      [FeeDataKey.Average]: {
+        txFee: String(averageFee),
+        chainSpecific: {
+          satoshiPerByte: averagePerByte
+        }
+      },
+      [FeeDataKey.Slow]: {
+        txFee: String(slowFee),
+        chainSpecific: {
+          satoshiPerByte: slowPerByte
+        }
+      }
+    } as FeeDataEstimate<T>
+  }
+
+  async signTransaction(signTxInput: SignTxInput<ChainTxType<T>>): Promise<string> {
+    try {
+      const { txToSign, wallet } = signTxInput
+      if (!supportsBTC(wallet))
+        throw new Error('UTXOBaseAdapter: signTransaction wallet does not support signing btc txs')
+
+      const signedTx = await wallet.btcSignTx(txToSign)
+      if (!signedTx) throw ErrorHandler('UTXOBaseAdapter: error signing tx')
+      return signedTx.serializedTx
+    } catch (err) {
+      return ErrorHandler(err)
+    }
   }
 
   async getTxHistory(input: TxHistoryInput): Promise<TxHistoryResponse<T>> {
@@ -253,6 +427,74 @@ export abstract class UTXOBaseAdapter<T extends UTXOChainIds> implements IChainA
       sendTxBody: { hex }
     })
     return data
+  }
+
+  async subscribeTxs(
+    input: SubscribeTxsInput,
+    onMessage: (msg: Transaction<T>) => void,
+    onError: (err: SubscribeError) => void
+  ): Promise<void> {
+    const { wallet } = input
+
+    let { bip44Params, accountType } = input
+    if (!bip44Params) {
+      bip44Params = this.getDefaultBip44Params()
+    }
+
+    if (!accountType) {
+      accountType = this.getDefaultAccountType()
+    }
+
+    const { xpub } = await this.getPublicKey(wallet, bip44Params, accountType)
+    const account = await this.getAccount(xpub)
+    const addresses = (account.chainSpecific.addresses ?? []).map((address) => address.pubkey)
+    const subscriptionId = `${toRootDerivationPath(bip44Params)}/${accountType}`
+
+    await this.providers.ws.subscribeTxs(
+      subscriptionId,
+      { topic: 'txs', addresses },
+      async (msg) => {
+        const tx = await this.getTransactionParser().parse(msg.data, msg.address)
+
+        onMessage({
+          address: tx.address,
+          blockHash: tx.blockHash,
+          blockHeight: tx.blockHeight,
+          blockTime: tx.blockTime,
+          chainId: tx.chainId,
+          chain: this.getType(),
+          confirmations: tx.confirmations,
+          fee: tx.fee,
+          status: getStatus(tx.status),
+          tradeDetails: tx.trade,
+          transfers: tx.transfers.map((transfer) => ({
+            assetId: transfer.assetId,
+            from: transfer.from,
+            to: transfer.to,
+            type: getType(transfer.type),
+            value: transfer.totalValue
+          })),
+          txid: tx.txid
+        })
+      },
+      (err) => onError({ message: err.message })
+    )
+  }
+
+  unsubscribeTxs(input?: SubscribeTxsInput): void {
+    if (!input) return this.providers.ws.unsubscribeTxs()
+
+    let { bip44Params, accountType } = input
+    if (!bip44Params) {
+      bip44Params = this.getDefaultBip44Params()
+    }
+
+    if (!accountType) {
+      accountType = this.getDefaultAccountType()
+    }
+    const subscriptionId = `${toRootDerivationPath(bip44Params)}/${accountType}`
+
+    this.providers.ws.unsubscribeTxs(subscriptionId, { topic: 'txs', addresses: [] })
   }
 
   async validateAddress(address: string): Promise<ValidAddressResult> {
